+++ seed_salesforce/salesforce_client.py
@@ -129,13 +129,15 @@ class SalesforceClient:
         else:
             raise Exception("Failed to return a Benchmark")
 
-    def get_benchmark_by_custom_id(self, salesforce_benchmark_id: str) -> dict:
-        """Return the benchmark by the Salesforce Benchmark ID.
+    def get_benchmark_by_custom_id(self, custom_id_name: str, custom_id_value: str) -> dict:
+        """Return the benchmark by the custom Salesforce Benchmark ID.
 
         Args:
-            salesforce_benchmark_id (str): Salesforce Benchmark ID of the property to return
-            Note: this is not necessarily the Benchmark ID (it's a separate field)
-            # TODO: make this configurable?
+            custom_id_name (str): Name of the custom ID field
+            custom_id_value (str): Value of the custom ID field
+            Note: this is not necessarily the Benchmark ID (it's a separate field that should be unique
+            and serves as the "key" for retrieving the benchmark, but it is not necessarily the actual
+            record ID in Salesforce)
 
         Returns:
             dict:  OrderedDict([('attributes',
@@ -146,7 +148,7 @@ class SalesforceClient:
         """
 
         benchmark_exist = self.connection.query(
-            format_soql("Select Id from Benchmark__c where Salesforce_Benchmark_ID__c = {}", salesforce_benchmark_id),
+            format_soql(f"Select Id from Benchmark__c where {custom_id_name} = {{}}", custom_id_value),
         )
         if len(benchmark_exist["records"]) == 1:
             # if there is a single record, then it exist, but
@@ -156,7 +158,7 @@ class SalesforceClient:
         elif len(benchmark_exist["records"]) > 1:
             # there are multiple properties with the same name, raise error
             raise Exception(
-                f"Failed to return Benchmark {salesforce_benchmark_id}...multiple benchmarks with that name found",
+                f"Failed to return Benchmark {custom_id_value}...multiple benchmarks with that name found",
             )
         else:
             # there is no property, return empty dict
@@ -208,6 +210,7 @@ class SalesforceClient:
             # return benchmark
             return updated_record
         else:
+
             raise Exception(
                 f"Failed to update Benchmark {salesforce_benchmark_id} with error: {updated_record['errors']}",
             )
@@ -306,6 +309,26 @@ class SalesforceClient:
             return self.connection.Property__c.get(property_id)
         except BaseException:
             raise Exception("Error retrieving property by ID")
+
+    def get_accounts(self) -> list:
+        """Get all accounts in salesforce
+
+        Returns:
+            list: list of accounts in salesforce
+        """
+        soql_query = "SELECT Id, Name FROM Account"
+        results = self.connection.query_all(soql_query)
+        return results['records']
+
+    def get_contacts(self) -> list:
+        """Get all contacts in salesforce
+
+        Returns:
+            list: list of contacts in salesforce
+        """
+        soql_query = "SELECT Id, Name, Email FROM Contact"
+        results = self.connection.query_all(soql_query)
+        return results['records']
 
     def get_account_by_account_id(self, account_id: str) -> dict:
         """Return the account by the account ID.