+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## Version 0.2.0
+Added a method to look up a benchmark with a custom ID name / value pair
+Updated testing to first retrieve a benchmark by custom ID and then update it
+Dependency updates
+
 ## Version 0.1.1
 
 Updated Python compatibility for v3.9-v3.12, and reformatted with Ruff